+++ Sources/OrmTxcSql/OrmTxcSql.csproj
@@ -53,9 +53,9 @@
     <Compile Include="Data\IParameterValueConverter.cs" />
     <Compile Include="Entities\AbstractEntity.cs" />
     <Compile Include="Entities\AbstractEntityExtensions.cs" />
-    <Compile Include="Extensions\DataTableExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Data\TooManyRowsException.cs" />
+    <Compile Include="Utils\DataTableExtensions.cs" />
     <Compile Include="Utils\LogUtils.cs" />
     <Compile Include="Utils\EntityUtils.cs" />
     <Compile Include="Utils\MessageUtils.cs" />
